+++ packages/lib-react-components/src/CloseButton/CloseButton.spec.js
@@ -17,4 +17,12 @@ describe('<CloseButton />', function () {
     wrapper.simulate('click')
     expect(wrapper.props().onClick).to.have.been.calledOnce()
   })
+
+  describe('with a color prop', function () {
+    it('should set the icon colour', function () {
+      const colouredButton = shallow(<CloseButton color='neutral-6' closeFn={sinon.spy()} />)
+      const icon = colouredButton.prop('icon')
+      expect(icon.props.color).to.equal('neutral-6')
+    })
+  })
 })